refactor: move definitions into `slot` module
